closes #30
