+++ lib/animated_toggle_switch.dart
@@ -1,5 +1,6 @@
 library animated_toggle_switch;
 
+import 'dart:async';
 import 'dart:math';
 
 import 'package:flutter/cupertino.dart';
@@ -8,6 +9,8 @@ import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 
 part 'src/properties.dart';
+part 'src/style.dart';
 part 'src/widgets/animated_toggle_switch.dart';
 part 'src/widgets/custom_animated_toggle_switch.dart';
 part 'src/widgets/drag_region.dart';
+part 'src/widgets/conditional_wrapper.dart';
